Layout for gramine-tyche
1. Adding a placeholder for Tyche

I copied the linux-sgx folder and adapted it to compile without the sgx
driver. I relied on a placeholder.h file to provide the missing
definitions. We will have to figure out how to replace the
implementation.

We can also use the skeleton folder to guide where we have a look.
We could have started from this one but let's try it this way first to
speed things up?

2. Infrastructure to install gramine-tyche

I modified the meson build to generate a gramine-tyche wrapper script
(similar to gramine-direct and gramine-sgx). There might still be some
adjustments to be made, but it seems to work on my machine and will
speed up testing and dev.

3. Adding the dependencies for Tyche

I managed to configure meson to import basically all the driver and sdk
libraries we need.
It mostly imports headers but add the source for the backend (which is
all we care about).