Fixing thread exiting
Due to TCS not being mapped to cores in SGX, they have a complex scheme
to notify the thread exited which is part of the thread reset of the
enclave.
In our case, we will do things slightly differently and just clear the
id upon exit, because we have pinned threads to cores.