Adding sandbox and quantum to sgx manifest
We might need to run domains as sandboxes (if we do not have
confidential VMs).
We also need to experiment with the quantum for scheduling to debug
performance.
I added two new toml entries in the manifest to tyche domains. They are
both available under sgx blocks e.g.:

[sgx]
quantum = 10
sandbox = true

The default values are sandbox = false and quantum = 10.
Quantum must be between 10 and 62, as it is a shift for the vmx timer.